--- src/app/(default)/album/analysis/_client-diagnosis-read.tsx
@@ -1,16 +1,0 @@
-"use client";
-
-import { useEffect } from "react";
-import { patchDiagnosisRead } from "@/lib/api/music-promotion";
-
-interface Props {
-  promotionId: number;
-}
-
-export default function ClientDiagnosisRead({ promotionId }: Props) {
-  useEffect(() => {
-    patchDiagnosisRead(promotionId);
-  }, [promotionId]);
-
-  return null;
-}
