+++ src/utils/date.ts
@@ -1,0 +1,10 @@
+export function formatDate(date: string | Date) {
+  return new Date(date)
+    .toLocaleDateString("ko-KR", {
+      year: "2-digit",
+      month: "2-digit",
+      day: "2-digit",
+    })
+    .replace(/\.\s/g, ".")
+    .replace(/\.$/, "");
+}
